+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Added an input for setting servings numbers for the cooking popup
+
 ## [0.0.2] - 2023-01-11
 
 ### Added